+++ djpcms/views/apps/docs.py
@@ -13,6 +13,7 @@ from djpcms.views.regex import RegExUrl
 
     
 class DocView(AppViewBase):
+    '''Sphinx documentation view.'''
     editurl          = None
     
     def __init__(self, regex = '(?P<url>[\w./-]*)', lang = False,
@@ -45,7 +46,7 @@ class DocView(AppViewBase):
         docroot = self.appmodel.get_docroot(lang, version)
     
         # First look for <bits>/index.fpickle, then for <bits>.fpickle
-        bits = url.strip('/').split('/') + ['index.fjson']
+        bits = url.strip('/').split('/') + ['%s.fjson' % app.master_doc]
         doc = docroot.child(*bits)
         if not doc.exists():
             bits = bits[:-2] + ['%s.fjson' % bits[-2]]
@@ -56,12 +57,18 @@ class DocView(AppViewBase):
         bits[-1] = bits[-1].replace('.fjson', '')
         name  = self.appmodel.name
         namet = '-'.join([b for b in bits if b])
+        namet = app.name_template_mapping.get(namet,namet)
         template_names = [
                           'docs/%s.html' % namet,
                           'djpcms/docs/%s.html' % namet,
                           'docs/doc.html',
                           ]
         doc = json.load(open(doc, 'rb'))
+        rels = []
+        for link in doc.get('rellinks',[]):
+            rels.append({'url': '%s%s/' % (app.baseurl,link[0]),
+                         'title': link[1]})
+        doc['rellinks'] = rels
         djp.breadcrumbs = self.makebreadcrumbs(djp,doc)
         c = {'doc':     doc,
              'env':     json.load(open(docroot.child('globalcontext.json'), 'rb')),
@@ -79,7 +86,7 @@ class DocView(AppViewBase):
         parent = djp
         if djp.urlargs:
             parent = parent.parent
-        parents = doc.get('parents',None)
+        parents = doc.get('parents',[])
         b = []
         while parent:
             b.append({'name': parent.title,
@@ -90,17 +97,28 @@ class DocView(AppViewBase):
             b.append({'url': parent.get('link',''),
                       'name': parent.get('title','')})
         if djp.urlargs:
-            b.append({'name': doc.get('title','')})
+            title = doc.get('title','')
+            if not title:
+                title = doc.get('indextitle','')
+            b.append({'name': title})
         for p in range(settings.ENABLE_BREADCRUMBS):
             b.pop(0)
         if b:
             b[-1].pop('url',None)
         return b
 
+
 class DocApplication(ApplicationBase):
     deflang          = 'en'
+    '''Default language. Default ``en``.'''
     defversion       = 'dev'
+    '''Default version. Default ``dev``.'''
     DOCS_PICKLE_ROOT = None
+    '''Root class for serialized documentation. Default ``None``.'''
+    master_doc       = 'index'
+    '''Specify sphinx master doc. Default ``index``.'''
+    name_template_mapping = {'py-modindex':'modindex'}
+    '''Dictionary which maps names to template names.'''
     
     index = DocView(regex = '')
     document = DocView(parent = 'index')
@@ -133,80 +151,3 @@ class DocApplication(ApplicationBase):
         css = {
             'all': ('djpcms/sphinx/smooth.css',)
         }
-    
-    
-class __OldDocApplication(ApplicationBase):
-        
-    def old__init__(self, baseurl, application_site, editavailable):
-        '''Create a tuple of urls'''
-        super(DocApplication,self).__init__(baseurl, application_site, editavailable)
-        urls = ()
-        appview                 = DocView(self)
-        self.root_application   = appview
-        self.language = None
-        self.version  = None
-        #self.urls = (url(r'^$', self.root_application),
-        #             url(r'^(?P<lang>[a-z-]+)/$', self.language),
-        #             url(r'^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/$',
-        #                 self.document, {'url': ''},  name = '%s_index' % self.name),
-        #             url(r'^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/(?P<url>[\w./-]*)/$',
-        #                 self.document, name = '%s_details' % self.name),
-        #            )
-        if self.deflang:
-            appview  = DocView(self,appview,lang=True)
-            self.language = appview
-            urls = url(r'^(?P<lang>[a-z-]+)/$',
-                       self.language,
-                       {'url': ''},
-                       name = '%s_lang' % self.name),
-        if self.defversion:
-            appview = DocView(self,appview,version=True)
-            self.version = appview
-            urls += url(r'^(?P<lang>[a-z-]+)/(?P<version>[\w.-]+)/$',
-                         self.version,
-                         {'url': ''},
-                         name = '%s_version' % self.name),
-        purl = ''
-        if urls:
-            urls = (url(r'^$', self.root_application),) + urls
-            if self.language:
-                purl = '(?P<lang>[a-z-]+)/'
-            if self.version:
-                purl += '(?P<version>[\w.-]+)/'
-        urls += url(r'^%s(?P<url>[\w./-]*)/$' % purl,
-                    appview,
-                    name = '%s_details' % self.name),
-        self.urls = urls
-            
-    
-    def get_root_code(self):
-        return self.name
-    
-    def get_path_args(self, lang, version):
-        return (lang, version, "_build", "json")
-    
-    def get_docroot(self, lang, version):
-        docroot = Path(self.DOCS_PICKLE_ROOT).child(*self.get_path_args(lang, version))
-        if not docroot.exists():
-            raise http.Http404()
-        return docroot 
-
-    def get_template(self,page):
-        '''
-        given a page objects (which may be None)
-        return the template file for the get view
-        '''
-        return None
-        
-    def bodybits(self):
-        if self.editurl:
-            return mark_safe(u'class="edit documentation"')
-        else:
-            return mark_safe(u'class="documentation"')
-        
-    def doc_index_url(self, request, lang, version):
-        return '%s%s/%s/' % (self.baseurl,lang,version)
-    
-    def table_of_content_url(self, request, lang, version):
-        return '%s%s/' % (self.doc_index(),'contents')
-